rft: make experimental modules
resolve #42
